Final energy price description
I propose to include in the description of final energy price, that this variable is a mean price. Final end users may buy their energy at different prices depending on the retailer they choose (or if they choose a regulated tariff when this exist). A unique price thus does not exist.